pyproject.toml
[project]
name = "icu-learn1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
  { name = "Masato Ota", email = "ohtam@nec.com" },
]
requires-python = ">=3.12"
dependencies = [
  "pyicu>=2.15.2",
]

[project.scripts]
icu-learn1 = "icu_learn1:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
